refactor: extract message-shape helpers
Eight pure readers over LangChain messages -- which messages belong to the
current turn, how to reach a tool result by call id, and how to coerce message
content to text -- move out of deepagents_runtime.py into message_shape.py.

They hold no runtime state and are the only cleanly separable boundary in the
file today. The larger diagnostics cut has a two-way dependency with the
evidence codecs and waits on that work.

All eight bodies are byte-identical to their originals; this is a move, not a
rewrite.

Note: _extract_final_text needs three skill_activation constants. An earlier
dependency analysis reported this leaf as having zero outward dependencies
because it counted only names defined in the module, not imported ones. The
constants import cleanly -- skill_activation does not import the runtime -- but
future extractions must count imported names too.
